pyproject.toml
[tool.poetry]
name = "new-grad"
version = "0.1.0"
description = ""
authors = ["Eren Torlak <erentorlak3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
gradio = "^5.9.1"
langchain = "^0.3.13"
langchain-chroma = "^0.1.4"
langsmith = "^0.2.4"
python-dotenv = "^1.0.1"
typing-extensions = "^4.12.2"
langchain-openai = "^0.2.14"
langchain-community = "^0.3.13"
langgraph = "^0.2.60"
bs4 = "^0.0.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
